std::sys::uefi: fix i16 overflow in to_uefi() year overflow path
The delta computation for year-10000 dates can reach 44,640 (day 31,
23:59:59) which exceeds i16::MAX (32,767), causing a silent wrap to a
negative value and triggering a panic at the subsequent assert.

Widen the intermediate computation to i32 and cast back to i16 only
after the range assert guarantees the value fits.

+++ library/std/src/sys/pal/uefi/system_time.rs
@@ -142,10 +142,10 @@ pub(crate) const fn to_uefi(dur: &Duration, timezone: i16, daylight: u8) -> Resu
         let delta = ((d - 1) as u64 * MIN_IN_DAY
             + hour as u64 * MIN_IN_HOUR
             + minute as u64
-            + if second == 0 { 0 } else { 1 }) as i16;
-        let new_tz = timezone + delta;
+            + if second == 0 { 0 } else { 1 }) as i32;
+        let new_tz = timezone as i32 + delta;
 
         assert!(new_tz <= 1440 && new_tz >= -1440);
-        Err(new_tz)
+        Err(new_tz as i16)
     }
 }
